Delete CSS paint code
