+++ lib/web_ui/lib/src/engine/surface/recording_canvas.dart
@@ -608,8 +608,6 @@ abstract class PaintCommand {
   const PaintCommand();
 
   void apply(EngineCanvas? canvas);
-
-  void serializeToCssPaint(List<List<dynamic>> serializedCommands);
 }
 
 /// A [PaintCommand] that affect pixels on the screen (unlike, for example, the
@@ -665,11 +663,6 @@ class PaintSave extends PaintCommand {
       return super.toString();
     }
   }
-
-  @override
-  void serializeToCssPaint(List<List<dynamic>> serializedCommands) {
-    serializedCommands.add(const <int>[1]);
-  }
 }
 
 class PaintRestore extends PaintCommand {
@@ -687,11 +680,6 @@ class PaintRestore extends PaintCommand {
     } else {
       return super.toString();
     }
-  }
-
-  @override
-  void serializeToCssPaint(List<List<dynamic>> serializedCommands) {
-    serializedCommands.add(const <int>[2]);
   }
 }
 
@@ -714,11 +702,6 @@ class PaintTranslate extends PaintCommand {
       return super.toString();
     }
   }
-
-  @override
-  void serializeToCssPaint(List<List<dynamic>> serializedCommands) {
-    serializedCommands.add(<num>[3, dx, dy]);
-  }
 }
 
 class PaintScale extends PaintCommand {
@@ -740,11 +723,6 @@ class PaintScale extends PaintCommand {
       return super.toString();
     }
   }
-
-  @override
-  void serializeToCssPaint(List<List<dynamic>> serializedCommands) {
-    serializedCommands.add(<num>[4, sx, sy]);
-  }
 }
 
 class PaintRotate extends PaintCommand {
@@ -765,11 +743,6 @@ class PaintRotate extends PaintCommand {
       return super.toString();
     }
   }
-
-  @override
-  void serializeToCssPaint(List<List<dynamic>> serializedCommands) {
-    serializedCommands.add(<num>[5, radians]);
-  }
 }
 
 class PaintTransform extends PaintCommand {
@@ -789,11 +762,6 @@ class PaintTransform extends PaintCommand {
     } else {
       return super.toString();
     }
-  }
-
-  @override
-  void serializeToCssPaint(List<List<dynamic>> serializedCommands) {
-    serializedCommands.add(<dynamic>[6]..addAll(matrix4));
   }
 }
 
@@ -816,11 +784,6 @@ class PaintSkew extends PaintCommand {
       return super.toString();
     }
   }
-
-  @override
-  void serializeToCssPaint(List<List<dynamic>> serializedCommands) {
-    serializedCommands.add(<num>[7, sx, sy]);
-  }
 }
 
 class PaintClipRect extends DrawCommand {
@@ -840,11 +803,6 @@ class PaintClipRect extends DrawCommand {
     } else {
       return super.toString();
     }
-  }
-
-  @override
-  void serializeToCssPaint(List<List<dynamic>> serializedCommands) {
-    serializedCommands.add(<dynamic>[8, _serializeRectToCssPaint(rect)]);
   }
 }
 
@@ -866,14 +824,6 @@ class PaintClipRRect extends DrawCommand {
       return super.toString();
     }
   }
-
-  @override
-  void serializeToCssPaint(List<List<dynamic>> serializedCommands) {
-    serializedCommands.add(<dynamic>[
-      9,
-      _serializeRRectToCssPaint(rrect),
-    ]);
-  }
 }
 
 class PaintClipPath extends DrawCommand {
@@ -893,11 +843,6 @@ class PaintClipPath extends DrawCommand {
     } else {
       return super.toString();
     }
-  }
-
-  @override
-  void serializeToCssPaint(List<List<dynamic>> serializedCommands) {
-    serializedCommands.add(<dynamic>[10, path.webOnlySerializeToCssPaint()]);
   }
 }
 
@@ -919,12 +864,6 @@ class PaintDrawColor extends DrawCommand {
     } else {
       return super.toString();
     }
-  }
-
-  @override
-  void serializeToCssPaint(List<List<dynamic>> serializedCommands) {
-    serializedCommands
-        .add(<dynamic>[11, colorToCssString(color), blendMode.index]);
   }
 }
 
@@ -948,18 +887,6 @@ class PaintDrawLine extends DrawCommand {
       return super.toString();
     }
   }
-
-  @override
-  void serializeToCssPaint(List<List<dynamic>> serializedCommands) {
-    serializedCommands.add(<dynamic>[
-      12,
-      p1.dx,
-      p1.dy,
-      p2.dx,
-      p2.dy,
-      _serializePaintToCssPaint(paint)
-    ]);
-  }
 }
 
 class PaintDrawPaint extends DrawCommand {
@@ -979,11 +906,6 @@ class PaintDrawPaint extends DrawCommand {
     } else {
       return super.toString();
     }
-  }
-
-  @override
-  void serializeToCssPaint(List<List<dynamic>> serializedCommands) {
-    serializedCommands.add(<dynamic>[13, _serializePaintToCssPaint(paint)]);
   }
 }
 
@@ -1006,11 +928,6 @@ class PaintDrawVertices extends DrawCommand {
       return super.toString();
     }
   }
-
-  @override
-  void serializeToCssPaint(List<List<dynamic>> serializedCommands) {
-    throw UnimplementedError();
-  }
 }
 
 class PaintDrawPoints extends DrawCommand {
@@ -1031,11 +948,6 @@ class PaintDrawPoints extends DrawCommand {
     } else {
       return super.toString();
     }
-  }
-
-  @override
-  void serializeToCssPaint(List<List<dynamic>> serializedCommands) {
-    throw UnimplementedError();
   }
 }
 
@@ -1058,15 +970,6 @@ class PaintDrawRect extends DrawCommand {
       return super.toString();
     }
   }
-
-  @override
-  void serializeToCssPaint(List<List<dynamic>> serializedCommands) {
-    serializedCommands.add(<dynamic>[
-      14,
-      _serializeRectToCssPaint(rect),
-      _serializePaintToCssPaint(paint)
-    ]);
-  }
 }
 
 class PaintDrawRRect extends DrawCommand {
@@ -1087,15 +990,6 @@ class PaintDrawRRect extends DrawCommand {
     } else {
       return super.toString();
     }
-  }
-
-  @override
-  void serializeToCssPaint(List<List<dynamic>> serializedCommands) {
-    serializedCommands.add(<dynamic>[
-      15,
-      _serializeRRectToCssPaint(rrect),
-      _serializePaintToCssPaint(paint),
-    ]);
   }
 }
 
@@ -1125,16 +1019,6 @@ class PaintDrawDRRect extends DrawCommand {
       return super.toString();
     }
   }
-
-  @override
-  void serializeToCssPaint(List<List<dynamic>> serializedCommands) {
-    serializedCommands.add(<dynamic>[
-      16,
-      _serializeRRectToCssPaint(outer),
-      _serializeRRectToCssPaint(inner),
-      _serializePaintToCssPaint(paint),
-    ]);
-  }
 }
 
 class PaintDrawOval extends DrawCommand {
@@ -1155,15 +1039,6 @@ class PaintDrawOval extends DrawCommand {
     } else {
       return super.toString();
     }
-  }
-
-  @override
-  void serializeToCssPaint(List<List<dynamic>> serializedCommands) {
-    serializedCommands.add(<dynamic>[
-      17,
-      _serializeRectToCssPaint(rect),
-      _serializePaintToCssPaint(paint),
-    ]);
   }
 }
 
@@ -1187,17 +1062,6 @@ class PaintDrawCircle extends DrawCommand {
       return super.toString();
     }
   }
-
-  @override
-  void serializeToCssPaint(List<List<dynamic>> serializedCommands) {
-    serializedCommands.add(<dynamic>[
-      18,
-      c.dx,
-      c.dy,
-      radius,
-      _serializePaintToCssPaint(paint),
-    ]);
-  }
 }
 
 class PaintDrawPath extends DrawCommand {
@@ -1218,15 +1082,6 @@ class PaintDrawPath extends DrawCommand {
     } else {
       return super.toString();
     }
-  }
-
-  @override
-  void serializeToCssPaint(List<List<dynamic>> serializedCommands) {
-    serializedCommands.add(<dynamic>[
-      19,
-      path.webOnlySerializeToCssPaint(),
-      _serializePaintToCssPaint(paint),
-    ]);
   }
 }
 
@@ -1252,22 +1107,6 @@ class PaintDrawShadow extends DrawCommand {
       return super.toString();
     }
   }
-
-  @override
-  void serializeToCssPaint(List<List<dynamic>> serializedCommands) {
-    serializedCommands.add(<dynamic>[
-      20,
-      path.webOnlySerializeToCssPaint(),
-      <dynamic>[
-        color.alpha,
-        color.red,
-        color.green,
-        color.blue,
-      ],
-      elevation,
-      transparentOccluder,
-    ]);
-  }
 }
 
 class PaintDrawImage extends DrawCommand {
@@ -1288,13 +1127,6 @@ class PaintDrawImage extends DrawCommand {
       return 'drawImage($image, $offset, $paint)';
     } else {
       return super.toString();
-    }
-  }
-
-  @override
-  void serializeToCssPaint(List<List<dynamic>> serializedCommands) {
-    if (assertionsEnabled) {
-      throw UnsupportedError('drawImage not serializable');
     }
   }
 }
@@ -1320,13 +1152,6 @@ class PaintDrawImageRect extends DrawCommand {
       return super.toString();
     }
   }
-
-  @override
-  void serializeToCssPaint(List<List<dynamic>> serializedCommands) {
-    if (assertionsEnabled) {
-      throw UnsupportedError('drawImageRect not serializable');
-    }
-  }
 }
 
 class PaintDrawParagraph extends DrawCommand {
@@ -1348,55 +1173,6 @@ class PaintDrawParagraph extends DrawCommand {
       return super.toString();
     }
   }
-
-  @override
-  void serializeToCssPaint(List<List<dynamic>> serializedCommands) {
-    if (assertionsEnabled) {
-      throw UnsupportedError('drawParagraph not serializable');
-    }
-  }
-}
-
-List<dynamic> _serializePaintToCssPaint(SurfacePaintData paint) {
-  final EngineGradient? engineShader = paint.shader as EngineGradient?;
-  return <dynamic>[
-    paint.blendMode?.index,
-    paint.style?.index,
-    paint.strokeWidth,
-    paint.strokeCap?.index,
-    paint.isAntiAlias,
-    colorToCssString(paint.color),
-    engineShader?.webOnlySerializeToCssPaint(),
-    paint.maskFilter?.webOnlySerializeToCssPaint(),
-    paint.filterQuality?.index,
-    paint.colorFilter?.webOnlySerializeToCssPaint(),
-  ];
-}
-
-List<dynamic> _serializeRectToCssPaint(ui.Rect rect) {
-  return <dynamic>[
-    rect.left,
-    rect.top,
-    rect.right,
-    rect.bottom,
-  ];
-}
-
-List<dynamic> _serializeRRectToCssPaint(ui.RRect rrect) {
-  return <dynamic>[
-    rrect.left,
-    rrect.top,
-    rrect.right,
-    rrect.bottom,
-    rrect.tlRadiusX,
-    rrect.tlRadiusY,
-    rrect.trRadiusX,
-    rrect.trRadiusY,
-    rrect.brRadiusX,
-    rrect.brRadiusY,
-    rrect.blRadiusX,
-    rrect.blRadiusY,
-  ];
 }
 
 class Subpath {
@@ -1421,14 +1197,6 @@ class Subpath {
     return result;
   }
 
-  List<dynamic> serializeToCssPaint() {
-    final List<dynamic> serialization = <dynamic>[];
-    for (int i = 0; i < commands.length; i++) {
-      serialization.add(commands[i].serializeToCssPaint());
-    }
-    return serialization;
-  }
-
   @override
   String toString() {
     if (assertionsEnabled) {
@@ -1439,25 +1207,10 @@ class Subpath {
   }
 }
 
-/// ! Houdini implementation relies on indices here. Keep in sync.
-class PathCommandTypes {
-  static const int moveTo = 0;
-  static const int lineTo = 1;
-  static const int ellipse = 2;
-  static const int close = 3;
-  static const int quadraticCurveTo = 4;
-  static const int bezierCurveTo = 5;
-  static const int rect = 6;
-  static const int rRect = 7;
-}
-
 abstract class PathCommand {
-  final int type;
-  const PathCommand(this.type);
+  const PathCommand();
 
   PathCommand shifted(ui.Offset offset);
-
-  List<dynamic> serializeToCssPaint();
 
   /// Transform the command and add to targetPath.
   void transform(Float32List matrix4, SurfacePath targetPath);
@@ -1472,16 +1225,11 @@ class MoveTo extends PathCommand {
   final double x;
   final double y;
 
-  const MoveTo(this.x, this.y) : super(PathCommandTypes.moveTo);
+  const MoveTo(this.x, this.y);
 
   @override
   MoveTo shifted(ui.Offset offset) {
     return MoveTo(x + offset.dx, y + offset.dy);
-  }
-
-  @override
-  List<dynamic> serializeToCssPaint() {
-    return <dynamic>[1, x, y];
   }
 
   @override
@@ -1504,16 +1252,11 @@ class LineTo extends PathCommand {
   final double x;
   final double y;
 
-  const LineTo(this.x, this.y) : super(PathCommandTypes.lineTo);
+  const LineTo(this.x, this.y);
 
   @override
   LineTo shifted(ui.Offset offset) {
     return LineTo(x + offset.dx, y + offset.dy);
-  }
-
-  @override
-  List<dynamic> serializeToCssPaint() {
-    return <dynamic>[2, x, y];
   }
 
   @override
@@ -1543,28 +1286,12 @@ class Ellipse extends PathCommand {
   final bool anticlockwise;
 
   const Ellipse(this.x, this.y, this.radiusX, this.radiusY, this.rotation,
-      this.startAngle, this.endAngle, this.anticlockwise)
-      : super(PathCommandTypes.ellipse);
+      this.startAngle, this.endAngle, this.anticlockwise);
 
   @override
   Ellipse shifted(ui.Offset offset) {
     return Ellipse(x + offset.dx, y + offset.dy, radiusX, radiusY, rotation,
         startAngle, endAngle, anticlockwise);
-  }
-
-  @override
-  List<dynamic> serializeToCssPaint() {
-    return <dynamic>[
-      3,
-      x,
-      y,
-      radiusX,
-      radiusY,
-      rotation,
-      startAngle,
-      endAngle,
-      anticlockwise,
-    ];
   }
 
   @override
@@ -1686,18 +1413,12 @@ class QuadraticCurveTo extends PathCommand {
   final double x2;
   final double y2;
 
-  const QuadraticCurveTo(this.x1, this.y1, this.x2, this.y2)
-      : super(PathCommandTypes.quadraticCurveTo);
+  const QuadraticCurveTo(this.x1, this.y1, this.x2, this.y2);
 
   @override
   QuadraticCurveTo shifted(ui.Offset offset) {
     return QuadraticCurveTo(
         x1 + offset.dx, y1 + offset.dy, x2 + offset.dx, y2 + offset.dy);
-  }
-
-  @override
-  List<dynamic> serializeToCssPaint() {
-    return <dynamic>[4, x1, y1, x2, y2];
   }
 
   @override
@@ -1734,18 +1455,12 @@ class BezierCurveTo extends PathCommand {
   final double x3;
   final double y3;
 
-  const BezierCurveTo(this.x1, this.y1, this.x2, this.y2, this.x3, this.y3)
-      : super(PathCommandTypes.bezierCurveTo);
+  const BezierCurveTo(this.x1, this.y1, this.x2, this.y2, this.x3, this.y3);
 
   @override
   BezierCurveTo shifted(ui.Offset offset) {
     return BezierCurveTo(x1 + offset.dx, y1 + offset.dy, x2 + offset.dx,
         y2 + offset.dy, x3 + offset.dx, y3 + offset.dy);
-  }
-
-  @override
-  List<dynamic> serializeToCssPaint() {
-    return <dynamic>[5, x1, y1, x2, y2, x3, y3];
   }
 
   @override
@@ -1782,8 +1497,7 @@ class RectCommand extends PathCommand {
   final double width;
   final double height;
 
-  const RectCommand(this.x, this.y, this.width, this.height)
-      : super(PathCommandTypes.rect);
+  const RectCommand(this.x, this.y, this.width, this.height);
 
   @override
   RectCommand shifted(ui.Offset offset) {
@@ -1825,11 +1539,6 @@ class RectCommand extends PathCommand {
   }
 
   @override
-  List<dynamic> serializeToCssPaint() {
-    return <dynamic>[6, x, y, width, height];
-  }
-
-  @override
   String toString() {
     if (assertionsEnabled) {
       return 'Rect($x, $y, $width, $height)';
@@ -1842,16 +1551,11 @@ class RectCommand extends PathCommand {
 class RRectCommand extends PathCommand {
   final ui.RRect rrect;
 
-  const RRectCommand(this.rrect) : super(PathCommandTypes.rRect);
+  const RRectCommand(this.rrect);
 
   @override
   RRectCommand shifted(ui.Offset offset) {
     return RRectCommand(rrect.shift(offset));
-  }
-
-  @override
-  List<dynamic> serializeToCssPaint() {
-    return <dynamic>[7, _serializeRRectToCssPaint(rrect)];
   }
 
   @override
@@ -1872,16 +1576,9 @@ class RRectCommand extends PathCommand {
 }
 
 class CloseCommand extends PathCommand {
-  const CloseCommand() : super(PathCommandTypes.close);
-
   @override
   CloseCommand shifted(ui.Offset offset) {
     return this;
-  }
-
-  @override
-  List<dynamic> serializeToCssPaint() {
-    return <dynamic>[8];
   }
 
   @override
