+++ lib/web_ui/test/golden_tests/engine/scuba.dart
@@ -91,7 +91,7 @@ typedef CanvasTest = FutureOr<void> Function(EngineCanvas canvas);
 
 /// Runs the given test [body] with each type of canvas.
 void testEachCanvas(String description, CanvasTest body,
-    {double maxDiffRate, bool bSkipHoudini = false}) {
+    {double maxDiffRate}) {
   const ui.Rect bounds = ui.Rect.fromLTWH(0, 0, 600, 800);
   test('$description (bitmap)', () {
     try {
@@ -123,18 +123,6 @@ void testEachCanvas(String description, CanvasTest body,
       TextMeasurementService.clearCache();
     }
   });
-  if (!bSkipHoudini) {
-    test('$description (houdini)', () {
-      try {
-        TextMeasurementService.initialize(rulerCacheCapacity: 2);
-        WebExperiments.instance.useCanvasText = false;
-        return body(HoudiniCanvas(bounds));
-      } finally {
-        WebExperiments.instance.useCanvasText = null;
-        TextMeasurementService.clearCache();
-      }
-    });
-  }
 }
 
 final ui.TextStyle _defaultTextStyle = ui.TextStyle(
